requirements.txt
lightrag-hku>=1.0.0
sentence-transformers>=3.0.0
pyyaml>=6.0
pytest>=8.0.0
